requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2.8
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
packaging==23.2
psycopg==3.1.16
psycopg-binary==3.1.16
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
rpds-py==0.16.2
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
uritemplate==4.1.1
